vibetube_backend/pyproject.toml
[project]
name = "vibetube-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[all]>=0.123.5",
    "imagekitio>=4.2.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
